requirements.txt
MinePI
Pillow
